FISH-8190 Remove server-base image
This makes the build complicated when using buildx due to the quirk that buildx cannot utilise the local image
cache - it can only pull from a published registry. The server-base Dockerfile is now just folded into the
other server dockerfiles. The downside to this is that the build for each image is now slower due to having to
redo all of what the server-base image previously did for each image (server-full jdk11, server-full jdk17,
server-full jdk21, server-web jdk11 etc.), and increases redundancy in the dockerfiles themselves.

Signed-off-by: Andrew Pielage <pandrex247@hotmail.com>