FISH-10814 Mark org.snmp4j.transport.unix as optional
The use of maven-bundle-plugin 5.1.9 instead of 4.1.0 (as used over in the Notifiers repo) has the repackaged SNMP4J attempt to import this package (even though by the maven dependency tree it should be the other way around as snmp4j-unix-transport depends on snmp4j?). I have marked this resolution as optional instead of pulling in this extra dependency as it requires Java 17.