FISH-12264 Remove redundant commands from ejb-invoker-secure tests
Each of the three tests in this module runs the set-ejb-invoker-configuration commands with different settings, so the initial call to enable the ejb-invoker is redundant as they all enable it anyway. If we want to have this test be configured via Maven then it will need refactoring into three separate examples or executions.

Signed-off-by: Andrew Pielage <pandrex247@hotmail.com>